+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/club/clubGround/JpaClubGroundRepository.java
@@ -1,6 +1,6 @@
-package com.mobble.mobbleserver.refactor.club.clubGround.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.club.clubGround;
 
-import com.mobble.mobbleserver.refactor.club.clubGround.entity.ClubGround;
+import com.mobble.mobbleserver.domain.club.clubGround.ClubGround;
 import com.mobble.mobbleserver.domain.ground.Ground;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ClubGroundRepository extends JpaRepository<ClubGround, Long> {
+public interface JpaClubGroundRepository extends JpaRepository<ClubGround, Long> {
 
     List<ClubGround> findByClubId(Long id);
 
@@ -16,5 +16,4 @@ public interface ClubGroundRepository extends JpaRepository<ClubGround, Long> {
 
     @Query("SELECT cg.ground FROM ClubGround cg WHERE cg.club.id = :clubId")
     List<Ground> findGroundsByClubId(@Param("clubId") Long clubId);
-
 }
